new migration

+++ models.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Integer, Column, String
+from sqlalchemy import BigInteger, Column, String
 from db import Base
 
 
 class Users(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    telegram_id = Column(BigInteger, unique=True)
     username = Column(String)
